+++ src/main/java/org/wise/vle/web/wise5/student/StudentPostDataController.java
@@ -1,27 +1,4 @@
-/**
- * Copyright (c) 2008-2017 Regents of the University of California (Regents).
- * Created by WISE, Graduate School of Education, University of California, Berkeley.
- *
- * This software is distributed under the GNU General Public License, v3,
- * or (at your option) any later version.
- *
- * Permission is hereby granted, without written agreement and without license
- * or royalty fees, to use, copy, modify, and distribute this software and its
- * documentation for any purpose, provided that the above copyright notice and
- * the following two paragraphs appear in all copies of this software.
- *
- * REGENTS SPECIFICALLY DISCLAIMS ANY WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
- * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE. THE SOFTWARE AND ACCOMPANYING DOCUMENTATION, IF ANY, PROVIDED
- * HEREUNDER IS PROVIDED "AS IS". REGENTS HAS NO OBLIGATION TO PROVIDE
- * MAINTENANCE, SUPPORT, UPDATES, ENHANCEMENTS, OR MODIFICATIONS.
- *
- * IN NO EVENT SHALL REGENTS BE LIABLE TO ANY PARTY FOR DIRECT, INDIRECT,
- * SPECIAL, INCIDENTAL, OR CONSEQUENTIAL DAMAGES, INCLUDING LOST PROFITS,
- * ARISING OUT OF THE USE OF THIS SOFTWARE AND ITS DOCUMENTATION, EVEN IF
- * REGENTS HAS BEEN ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package org.wise.vle.web.wise5;
+package org.wise.vle.web.wise5.student;
 
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -37,13 +14,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.wise.portal.dao.ObjectNotFoundException;
-import org.wise.portal.domain.authentication.impl.StudentUserDetails;
 import org.wise.portal.domain.project.impl.ProjectComponent;
 import org.wise.portal.domain.run.Run;
 import org.wise.portal.domain.user.User;
@@ -61,16 +34,9 @@ import org.wise.vle.domain.work.StudentWork;
 
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
-/**
- * Controller for handling GET and POST requests of WISE5 student data WISE5 student data is stored
- * as StudentWork, Event, Annotation, and StudentAsset domain objects
- *
- * @author Hiroki Terashima
- */
-@Secured("ROLE_STUDENT")
 @Controller
-@RequestMapping("/api/student/data")
-public class StudentDataController {
+@Secured("ROLE_STUDENT")
+public class StudentPostDataController {
 
   @Autowired
   private ProjectService projectService;
@@ -93,143 +59,8 @@ public class StudentDataController {
   @Autowired
   private BroadcastStudentWorkService broadcastStudentWorkService;
 
-  @GetMapping
-  public void getWISE5StudentData(HttpServletResponse response, Authentication authentication,
-      @RequestParam(value = "getStudentWork", defaultValue = "false") boolean getStudentWork,
-      @RequestParam(value = "getEvents", defaultValue = "false") boolean getEvents,
-      @RequestParam(value = "getAnnotations", defaultValue = "false") boolean getAnnotations,
-      @RequestParam(value = "id", required = false) Integer id,
-      @RequestParam(value = "runId", required = false) Integer runId,
-      @RequestParam(value = "periodId", required = false) Integer periodId,
-      @RequestParam(value = "workgroupId", required = false) Integer workgroupId,
-      @RequestParam(value = "isAutoSave", required = false) Boolean isAutoSave,
-      @RequestParam(value = "isSubmit", required = false) Boolean isSubmit,
-      @RequestParam(value = "nodeId", required = false) String nodeId,
-      @RequestParam(value = "componentId", required = false) String componentId,
-      @RequestParam(value = "componentType", required = false) String componentType,
-      @RequestParam(value = "context", required = false) String context,
-      @RequestParam(value = "category", required = false) String category,
-      @RequestParam(value = "event", required = false) String event,
-      @RequestParam(value = "fromWorkgroupId", required = false) Integer fromWorkgroupId,
-      @RequestParam(value = "toWorkgroupId", required = false) Integer toWorkgroupId,
-      @RequestParam(value = "studentWorkId", required = false) Integer studentWorkId,
-      @RequestParam(value = "localNotebookItemId", required = false) String localNotebookItemId,
-      @RequestParam(value = "notebookItemId", required = false) Integer notebookItemId,
-      @RequestParam(value = "annotationType", required = false) String annotationType,
-      @RequestParam(value = "components", required = false) List<JSONObject> components,
-      @RequestParam(value = "onlyGetLatest", required = false) Boolean onlyGetLatest)
-      throws ObjectNotFoundException, IOException, JSONException {
-    JSONObject result = new JSONObject();
-    User user = userService.retrieveUser((StudentUserDetails) authentication.getPrincipal());
-    Run run = runService.retrieveById(Long.valueOf(runId));
-    if (getStudentWork && isMemberOfWorkgroupId(user, run, workgroupId)) {
-      try {
-        result.put("studentWorkList", getStudentWork(id, runId, periodId, workgroupId,
-            isAutoSave, isSubmit, nodeId, componentId, componentType, components, onlyGetLatest));
-      } catch (JSONException e) {
-        e.printStackTrace();
-      }
-    }
-    if (getEvents && isMemberOfWorkgroupId(user, run, workgroupId)) {
-      try {
-        result.put("events", getEvents(id, runId, periodId, workgroupId, nodeId, componentId,
-            componentType, context, category, event, components));
-      } catch (JSONException e) {
-        e.printStackTrace();
-      }
-    }
-    if (getAnnotations && isAllowedToGetAnnotations(user, run, fromWorkgroupId, toWorkgroupId)) {
-      try {
-        result.put("annotations", getAnnotations(id, runId, periodId, fromWorkgroupId,
-            toWorkgroupId, nodeId, componentId, studentWorkId, localNotebookItemId,
-            notebookItemId, annotationType));
-      } catch (JSONException e) {
-        e.printStackTrace();
-      }
-    }
-    try {
-      PrintWriter writer = response.getWriter();
-      writer.write(result.toString());
-      writer.close();
-    } catch (IOException e) {
-      e.printStackTrace();
-    }
-  }
-
-  private boolean isAllowedToGetAnnotations(User user, Run run, Integer fromWorkgroupId,
-      Integer toWorkgroupId) throws ObjectNotFoundException {
-    return isMemberOfWorkgroupId(user, run, fromWorkgroupId) ||
-        isMemberOfWorkgroupId(user, run, toWorkgroupId);
-  }
-
-  private boolean isMemberOfWorkgroupId(User user, Run run, Integer workgroupId)
-      throws ObjectNotFoundException {
-    return workgroupId != null && workgroupService.isUserInWorkgroupForRun(user, run,
-        workgroupService.retrieveById(Long.valueOf(workgroupId)));
-  }
-
-  private JSONArray getStudentWork(Integer id, Integer runId, Integer periodId, Integer workgroupId,
-      Boolean isAutoSave, Boolean isSubmit, String nodeId, String componentId, String componentType,
-      List<JSONObject> components, Boolean onlyGetLatest) {
-    List<StudentWork> studentWorkList = vleService.getStudentWorkList(id, runId, periodId,
-        workgroupId, isAutoSave, isSubmit, nodeId, componentId, componentType, components,
-        onlyGetLatest);
-    JSONArray studentWorkJSONArray = new JSONArray();
-    for (StudentWork studentWork : studentWorkList) {
-      studentWorkJSONArray.put(studentWork.toJSON());
-    }
-    return studentWorkJSONArray;
-  }
-
-  private JSONArray getEvents(Integer id, Integer runId, Integer periodId,
-    Integer workgroupId, String nodeId, String componentId, String componentType, String context,
-    String category, String event, List<JSONObject> components) {
-    List<Event> events = vleService.getEvents(id, runId, periodId, workgroupId, nodeId,
-        componentId, componentType, context, category, event, components);
-    JSONArray eventsJSONArray = new JSONArray();
-    for (Event eventObject : events) {
-      eventsJSONArray.put(eventObject.toJSON());
-    }
-    return eventsJSONArray;
-  }
-
-  private JSONArray getAnnotations(Integer id, Integer runId, Integer periodId,
-      Integer fromWorkgroupId, Integer toWorkgroupId, String nodeId, String componentId,
-      Integer studentWorkId, String localNotebookItemId, Integer notebookItemId,
-      String annotationType) {
-    List<Annotation> annotations = vleService.getAnnotations(id, runId, periodId, fromWorkgroupId,
-        toWorkgroupId, nodeId, componentId, studentWorkId, localNotebookItemId, notebookItemId,
-        annotationType);
-    JSONArray annotationsJSONArray = new JSONArray();
-    for (Annotation annotation : annotations) {
-      annotationsJSONArray.put(annotation.toJSON());
-    }
-    return annotationsJSONArray;
-  }
-
-  public void broadcastAnnotationToTeacher(Annotation annotation) throws JSONException {
-    JSONObject message = new JSONObject();
-    message.put("type", "annotationToTeacher");
-    message.put("topic", String.format("/topic/teacher/%s", annotation.getRunId()));
-    message.put("annotation", annotation.toJSON());
-    redisPublisher.publish(message.toString());
-  }
-
-  /**
-   * Handles batch POSTing student data (StudentWork, Action, Annotation)
-   *
-   * @param runId
-   *                          Run that the POSTer (student) is in
-   * @param studentWorkList
-   *                          JSON string containing student work, ex:
-   *                          [{"runId":2,"nodeId":"node4",...},{"runId":2,"nodeId":"node5",...}]
-   * @param events
-   *                          JSON string containing events
-   * @param annotations
-   *                          JSON string containing annotations
-   */
-  @PostMapping
-  public void postWISE5StudentData(HttpServletResponse response,
+  @PostMapping("/api/student/data")
+  public void postStudentData(HttpServletResponse response,
       @RequestBody ObjectNode postedParams, Authentication auth) throws JSONException {
     User user = userService.retrieveUserByUsername(auth.getName());
     Integer runId = postedParams.get("runId").asInt();
@@ -520,5 +351,13 @@ public class StudentDataController {
     } catch (ObjectNotFoundException | JSONException e) {
     }
     return false;
+  }
+
+  private void broadcastAnnotationToTeacher(Annotation annotation) throws JSONException {
+    JSONObject message = new JSONObject();
+    message.put("type", "annotationToTeacher");
+    message.put("topic", String.format("/topic/teacher/%s", annotation.getRunId()));
+    message.put("annotation", annotation.toJSON());
+    redisPublisher.publish(message.toString());
   }
 }
